+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   `verifyTokenExpirationOnTabFocus` option
+-   `logger` option
 
 ## 3.5.2 - 2019-03-16